advise against using pass file (SOFTWARE-4086)
the directions failed to mention that the password file would have to
be sourced ... but i'm not wild about it in general anyway so i'll
recommend something else.  Though for most cases if you are tagging
a single image, letting the script read your creds directly is fine,
and maybe the most straightforward.